daily_shorts/requirements.txt
requests>=2.31.0
moviepy>=1.0.3
Pillow>=9.5.0
python-dotenv>=1.0.0
numpy>=1.24.0
